+++ src/vue-router/routes.js
@@ -5,4 +5,5 @@ export const vueRoutes = Object.freeze({
   profile: { name: 'app.profile' },
   willRequests: { name: 'app.will-requests' },
   willRequestsList: { name: 'app.will-requests.list' },
+  willRequestsDetails: { name: 'app.will-requests.details' },
 })