+++ src/vue/pages/WillRequests/WillRequestsDetails.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="will-requests-details">
+    <!--  -->
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'will-requests-details',
+}
+</script>